requirements-selfbot.txt
discord.py-self[voice]>=2.0.0
